# api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Raikada Recording Server API
  description: |
    HTTP control API exposed by the Raikada Recording Server — the
    on-prem appliance defined in this repository. This is the recorder's
    local control surface. It is invoked by the Management Server over
    the recorder-initiated WebSocket per ADR 0003 (Connectivity model),
    and by LAN clients when authorized. Off-LAN clients reach this
    surface via the Cloud-brokered FRP tunnel.

    **/v1 reshape (ADR 0009).** This document describes the canonical
    `/v1/...` API per ADR 0009. The MediaMTX-lineage `/v3` surface is
    gone; there is no deprecation window and no alias. Every endpoint
    keys on canonical UUIDs (per ADR 0009 §D4); every request and
    response shape uses **snake_case** JSON keys. The single small
    legacy carve-out is `Stream.tracks[].codec_props`: those nine
    discriminated codec sub-shapes inherit the upstream MediaMTX
    camelCase JSON tags (`channelCount`, `sampleRate`, `bitDepth`,
    `muLaw`) and have not been renamed; the inconsistency is tracked
    as a Phase 3 follow-up.

    **In-house error envelope (D15).** Errors return the in-house
    `{ status: "error", error: <message> }` shape, not RFC 7807
    `application/problem+json`. Successful writes that have nothing to
    say return `{ status: "ok" }`. The decision to keep the in-house
    shape is documented in canonical-divergence D15 and a future
...

...

...

...

...
      with them.

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
  - bearerJWT: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        HTTP Basic credentials. Validated against the recorder's
        configured auth manager (internal, HTTP, or JWT-backed) on
        every request.
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Standard JWT bearer token. Validated against JWKS held by the
        recorder's auth manager. Use
        `POST /v1/auth/refresh-issuer-material` to force a JWKS
        reload after key rotation.

        Note: a `Bearer` token that happens to contain a colon (e.g.
        the upstream MediaMTX `user:pass` convention) is treated
        opaquely as a token here, not as colon-separated credentials.
        Clients that need username+password authentication must use
        `basicAuth`. See D6 in
        `recorder/docs/canonical-divergences.md`.
...
paths:
...
  /v1/info:
    get:
      tags: [Info]
      summary: Recorder runtime info.
      description: |
        Returns recorder-self-reported version, start time, and the
        bound `tenant_id`. Same role as the legacy `/v3/info`; the
        URL prefix change is the only difference.
      responses:
        '200':
          description: Info response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== Auth =====================================================
  /v1/auth/refresh-issuer-material:
    post:
      tags: [Auth]
      summary: Force the recorder's auth manager to refresh cached issuer material.
      description: |
        Renamed mechanism-neutrally from the legacy
        `/v3/auth/jwks/refresh` per ADR 0009 §D7. ADR 0011 picked
        JWT/JWKS for user-facing flows and mTLS X.509 for service-to-
        service connections; the mechanism-neutral name was retained
        so additional issuer-material kinds (e.g., the trust roots
        for mTLS validation) can flow through this endpoint without
        a rename. Today the implementation refreshes the JWKS cache.
      responses:
        '200':
          description: Refresh accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== Cameras ==================================================
  /v1/cameras:
    get:
      tags: [Cameras]
      summary: List cameras.
      description: |
        Returns canonical Camera entities including the small
        `runtime` block per ADR 0009 §D3. Sorted deterministically
        by recorder-internal path-name (UUID round-trip stable).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ItemsPerPage'
      responses:
        '200':
          description: Camera list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraList'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Cameras]
      summary: Create a camera.
      description: |
        Server issues the canonical UUID per ADR 0009 §D4. The
        `name` field is required. The `tenant_id` field on the body
        is optional; if present and mismatched against the
        recorder's bound tenant, the recorder returns 403.
        Server-managed fields (`id`, `runtime`, `created_at`,
        `updated_at`) on the body are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Camera'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Camera with that name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/cameras/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Cameras]
      summary: Get a camera by canonical UUID.
      responses:
        '200':
          description: Camera.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          description: Invalid camera id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Camera not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Cameras]
      summary: Partially update a camera.
      description: |
        Partial-update semantics: only fields present in the body are
        applied. The `runtime` block is read-only (writes are
        ignored). Body's `tenant_id`, if present, must match the
        recorder's bound tenant or the request returns 403.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Camera'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Camera not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Cameras]
      summary: Replace a camera.
      description: |
        Full-replace semantics: every conf-side field is overwritten
        from the body. Server-managed fields (`id`, `runtime`,
        `created_at`) are preserved. Body's `tenant_id`, if present,
        must match.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Camera'
      responses:
        '200':
          description: Replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Camera'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Camera not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Cameras]
      summary: Delete a camera.
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: Invalid camera id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Camera not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== RecordingPolicies =======================================
  /v1/recording-policies:
    get:
      tags: [RecordingPolicies]
      summary: List recording policies.
      description: |
        Returns canonical RecordingPolicy entries. Pre-MS the
        recorder lazily synthesizes policies from per-camera
        recording fields on first access; per-restart persistence is
        intentional (synthesis re-runs on next start).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ItemsPerPage'
      responses:
        '200':
          description: Policy list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPolicyList'
    post:
      tags: [RecordingPolicies]
      summary: Create a recording policy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordingPolicy'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPolicy'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/recording-policies/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [RecordingPolicies]
      summary: Get a recording policy.
      responses:
        '200':
          description: Policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPolicy'
        '400':
          description: Invalid id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [RecordingPolicies]
      summary: Partially update a recording policy.
      description: |
        When the patch flips `enabled`, every camera that references
        this policy has its underlying recording state toggled
        accordingly; the change is applied immediately to the
        recorder's per-path `Record` flag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordingPolicy'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPolicy'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Tenant mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [RecordingPolicies]
      summary: Delete a recording policy.
      description: |
        Returns 409 Conflict if any Camera still references the
        policy via `Camera.recording_policy_id`. Per ADR 0009 §D5.
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Policy still referenced by one or more cameras.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== Streams ==================================================
  /v1/streams:
    get:
      tags: [Streams]
      summary: List active runtime streams.
      description: |
        Folds 25 protocol-specific endpoints from the legacy /v3
        surface into a single canonical surface per ADR 0009 §D5.
        Per-protocol detail lives on the response in
        `protocol_specific`. Filters are AND-combined.

        `remote_addr` and `query_string` are redacted on output per
        canonical-divergences D5 / D12.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ItemsPerPage'
        - name: protocol
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StreamProtocol'
        - name: camera_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: direction
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StreamDirection'
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/StreamState'
      responses:
        '200':
          description: Stream list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamList'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/streams/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Streams]
      summary: Get a single active stream by canonical UUID.
      description: |
        Searches every protocol cluster for the UUID; first match
        wins. The response's `protocol_specific` payload is keyed by
        the matched cluster.
      responses:
        '200':
          description: Stream detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stream'
        '400':
          description: Invalid stream id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Stream not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Streams]
      summary: Terminate (kick) an active stream.
      description: |
        DELETE on a Stream means **terminate the session**, not
        delete a stored record — the resource is ephemeral, and
        DELETE is the canonical kick verb for a runtime resource.
      responses:
        '200':
          description: Kicked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: Invalid stream id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Stream not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== Recordings ===============================================
  /v1/recordings:
    get:
      tags: [Recordings]
      summary: List recordings.
      description: |
        Returns canonical Recording entities. List responses **omit**
        the embedded `segments` array per ADR 0009 §D5 — fetch a
        single recording for the segment list.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/ItemsPerPage'
        - name: camera_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RecordingState'
        - name: started_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: started_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Recording list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingList'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/recordings/{id}/playback:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Recordings]
      summary: Issue a playback handle for a recording.
      description: |
        Returns playback connection material (URL + `expires_at`
        validity bound). The actual stream comes through the existing
        media protocols, not over this endpoint. Per ADR 0011 the
        recorder does not issue tokens — clients reuse the user JWT
        they already hold from Cloud/MS when fetching the URL; the
        playback `/get` endpoint authenticates via the existing
        `auth.Manager`.
      responses:
        '200':
          description: Playback handle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackHandle'
        '400':
          description: Invalid recording id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Recording not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== RecordingSegments =======================================
...
  /v1/clips/{id}/download:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Clips]
      summary: Download the stitched clip bytes.
      description: |
        Returns the stitched export artifact with
        `Content-Type: video/mp4`. The clip must be in `ready`
        state; other states return 409.

        Authentication: per ADR 0011 the recorder does not issue
        per-download tokens; this endpoint is gated by the same JWT
        validation middleware as the rest of the `/v1` surface.
      responses:
        '200':
          description: Stitched clip bytes.
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid clip id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Clip not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Clip is not in `ready` state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== Events ===================================================
...